Feature ETP-2352: Move methods to util
